Fix bugs and improve logging
Fixed final evaluation by ensuring best model was being saved in a
checkpoint and was being loaded for evaluation. Also fixed a bug where
the final set of inputs wasn't guaranteed to fill the model input since
total number of inputs may not have been divisible by stride length.

Finally, added logging of the number of model parameters and the number
of training and evaluation examples.